converter.py: Leave "quoted" text unconverted.
Note: the quoted text still passes through the standardize_beta_code()
function first, which has the potential to mutilate the text.

This doesn't occur in practice, because the quoted text doesn't contain any
upper case or breathing marks.  But if it did, some restructuring would be
needed to only standardize the text that is converted.